crypto/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-crontab==0.7.1
django-environ==0.9.0
djangorestframework==3.14.0
idna==3.4
influxdb-client==1.33.0
protobuf==4.21.4
psycopg2-binary==2.9.4
python-dateutil==2.8.2
pytz==2022.4
reactivex==4.0.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tabulate==0.8.10
typing_extensions==4.4.0
urllib3==1.26.12
